src/backend/requirements.txt
anyio==3.7.0
asgiref==3.7.2
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
cleanurl==0.1.15
cryptography==41.0.1
defusedxml==0.7.1
Deprecated==1.2.14
Django==4.2.2
django-allauth==0.54.0
django-braces==1.15.0
django-guardian==2.4.0
django-oauth-toolkit==2.3.0
django-rest-auth==0.9.5
django-rest-framework-social-oauth2==1.1.0
djangorestframework==3.14.0
greenlet==2.0.2
h11==0.14.0
httpcore==0.17.2
httpx==0.24.1
idna==3.4
iniconfig==2.0.0
jsonify==0.5
jwcrypto==1.5.0
langcodes==3.3.0
oauthlib==3.2.2
packaging==23.1
playwright==1.35.0
pluggy==1.2.0
pycparser==2.21
pyee==9.0.4
PyJWT==2.7.0
pytest==7.4.0
pytest-base-url==2.0.0
pytest-playwright==0.3.3
python-slugify==8.0.1
python3-openid==3.2.0
pytz==2023.3
requests==2.31.0
requests-oauthlib==1.3.1
robotexclusionrulesparser==1.7.1
selectolax==0.3.14
six==1.16.0
sniffio==1.3.0
social-auth-app-django==5.2.0
social-auth-core==4.4.2
sqlparse==0.4.4
text-unidecode==1.3
typing_extensions==4.7.0
urllib3==2.0.3
wrapt==1.15.0
